fixed delete method in ProductService

+++ src/main/java/my/flick/rd/hw3/service/impl/ProductServiceImpl.java
@@ -63,7 +63,9 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void deleteProduct(long id) {
-        productRepository.deleteById(id);
+        if(productRepository.existsById(id)) {
+            productRepository.deleteById(id);
+        }
     }
 
     //Todo: consider moving to another class
